elas-erp/backend/requirements.txt
# Web API
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Core utils
python-multipart==0.0.9
pydantic==2.12.4
pydantic-settings==2.6.1

# DB
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
supabase==2.23.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Caching/queues
redis==5.0.8

# Storage
boto3==1.34.162

# Data wrangling / profiling
pandas==2.2.2
duckdb==1.1.3

# Document processing
PyPDF2==3.0.1
openpyxl==3.1.2
python-docx==1.1.0
pdfplumber==0.11.0

# HTTP requests
requests==2.31.0

# Agents / LLMs (Groq)
langchain==0.2.16
langchain-community==0.2.16
langgraph==0.2.33
langchain-groq==0.1.4
groq==0.11.0
tiktoken==0.7.0

# PII detection (optional)
presidio-analyzer==2.2.357
presidio-anonymizer==2.2.357

# File handling
aiofiles==24.1.0

# Validation/testing
pytest==8.3.3
httpx==0.27.2
